requirements.txt
# Data & Computation
numpy
pandas
scipy

# Financial Data
yfinance
pandas-datareader

# Technical Analysis
pandas-ta

# Visualization
matplotlib
plotly
seaborn

# Machine Learning
scikit-learn
xgboost

# Dashboard
streamlit

# Database
sqlalchemy

# Utilities
pyyaml
python-dateutil
tqdm
loguru

# Jupyter
jupyter
notebook

# Testing
pytest